export HTMLContentConnected parts
